# lib/api-zod/src/generated/api/api.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Bruce Football Performance Diagnostics — API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Kontaktformular absenden
 */
export const submitContactBodyNameMin = 2;

export const submitContactBodyMessageMin = 10;



export const SubmitContactBody = zod.object({
  "name": zod.string().min(submitContactBodyNameMin),
  "email": zod.string(),
  "subject": zod.string().nullish(),
  "message": zod.string().min(submitContactBodyMessageMin),
  "phone": zod.string().nullish(),
  "datenschutz": zod.boolean().optional()
})

export const SubmitContactResponse = zod.object({
  "success": zod.boolean(),
  "message": zod.string()
})


/**
 * @summary Demo-Anfrage / Interessent erfassen
 */
export const submitLeadBodyNameMin = 2;

export const submitLeadBodyVereinsnameMin = 2;



export const SubmitLeadBody = zod.object({
  "name": zod.string().min(submitLeadBodyNameMin),
  "email": zod.string(),
  "vereinsname": zod.string().min(submitLeadBodyVereinsnameMin),
  "telefon": zod.string().nullish(),
  "spieleranzahl": zod.string().nullish(),
  "nachricht": zod.string().nullish(),
  "plan": zod.string().nullish().describe('starter | professional | enterprise'),
  "datenschutz": zod.boolean().optional()
})

export const SubmitLeadResponse = zod.object({
  "success": zod.boolean(),
  "message": zod.string()
})


/**
 * @summary Gibt zurück ob die App im Wartungsmodus ist
 */
export const GetMaintenanceStatusResponse = zod.object({
  "maintenance": zod.boolean(),
  "message": zod.string().nullish()
})


/**
 * @summary Stripe Webhook Empfänger
 */
export const StripeWebhookBody = zod.object({
  "type": zod.string(),
  "id": zod.string().optional()
})

export const StripeWebhookResponse = zod.unknown()


/**
 * Authenticate a trainer with email and password
 * @summary Mobile login
 */
export const MobileLoginBody = zod.object({
  "email": zod.string(),
  "password": zod.string()
})

export const MobileLoginResponse = zod.object({
  "token": zod.string(),
  "user": zod.object({
  "id": zod.number(),
  "email": zod.string(),
  "vorname": zod.string(),
  "nachname": zod.string(),
  "rolle": zod.string(),
  "verein_id": zod.number().nullish(),
  "verein_name": zod.string().nullish()
})
})


/**
 * Returns players assigned to the authenticated trainer
 * @summary List players
 */
export const MobileGetPlayersResponse = zod.object({
  "players": zod.array(zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "vorname": zod.string(),
  "nachname": zod.string(),
  "mannschaft": zod.string().nullish(),
  "altersklasse": zod.string().nullish(),
  "score": zod.number().nullish(),
  "last_test_date": zod.string().nullish()
}))
})


/**
 * Returns full player profile with latest test results and athletik score
 * @summary Get player profile
 */
export const MobileGetPlayerParams = zod.object({
  "playerId": zod.coerce.number()
})

export const MobileGetPlayerResponse = zod.object({
  "player": zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "vorname": zod.string(),
  "nachname": zod.string(),
  "mannschaft": zod.string().nullish(),
  "altersklasse": zod.string().nullish(),
  "score": zod.number().nullish(),
  "last_test_date": zod.string().nullish()
}),
  "score": zod.number().nullish(),
  "tests": zod.object({
  "fms": zod.object({
  "datum": zod.string().optional(),
  "score": zod.number().optional(),
  "bewertung": zod.string().optional(),
  "asymmetrie": zod.string().nullish()
}).nullish(),
  "sprint": zod.object({
  "datum": zod.string().optional(),
  "beste_10m": zod.number().nullish(),
  "beste_30m": zod.number().nullish(),
  "bewertung_10m": zod.string().nullish()
}).nullish(),
  "ybalance": zod.object({
  "datum": zod.string().optional(),
  "composite_rechts": zod.number().optional(),
  "composite_links": zod.number().optional(),
  "asymmetrie": zod.string().nullish()
}).nullish()
})
})


/**
 * Save a new FMS test result for a player
 * @summary Submit FMS test
 */
export const MobileSubmitFmsParams = zod.object({
  "playerId": zod.coerce.number()
})

export const MobileSubmitFmsBody = zod.object({
  "datum": zod.string(),
  "deep_squat": zod.number(),
  "hurdle_l": zod.number(),
  "hurdle_r": zod.number(),
  "inline_l": zod.number(),
  "inline_r": zod.number(),
  "shoulder_l": zod.number(),
  "shoulder_r": zod.number(),
  "aslr_l": zod.number(),
  "aslr_r": zod.number(),
  "trunk": zod.number(),
  "rotary_l": zod.number(),
  "rotary_r": zod.number()
})

export const MobileSubmitFmsResponse = zod.object({
  "score": zod.number().nullish(),
  "sub_score": zod.number().nullish(),
  "message": zod.string()
})


/**
 * Save a new sprint test result for a player
 * @summary Submit sprint test
 */
export const MobileSubmitSprintParams = zod.object({
  "playerId": zod.coerce.number()
})

export const MobileSubmitSprintBody = zod.object({
  "datum": zod.string(),
  "best_10m": zod.number(),
  "best_30m": zod.number().optional()
})

export const MobileSubmitSprintResponse = zod.object({
  "score": zod.number().nullish(),
  "sub_score": zod.number().nullish(),
  "message": zod.string()
})


/**
 * Save a new Y-Balance test result for a player
 * @summary Submit Y-Balance test
 */
export const MobileSubmitYbalanceParams = zod.object({
  "playerId": zod.coerce.number()
})

export const MobileSubmitYbalanceBody = zod.object({
  "datum": zod.string(),
  "ant_r": zod.number(),
  "ant_l": zod.number(),
  "pm_r": zod.number(),
  "pm_l": zod.number(),
  "pl_r": zod.number(),
  "pl_l": zod.number(),
  "leg_length_r": zod.number().nullish(),
  "leg_length_l": zod.number().nullish()
})

export const MobileSubmitYbalanceResponse = zod.object({
  "score": zod.number().nullish(),
  "sub_score": zod.number().nullish(),
  "message": zod.string()
})